f1tenth/src/dcsl_f1tenth/dcsl_f1tenth/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
attrs==21.2.0
autoware-auto-geometry-msgs==1.0.0
autoware-auto-mapping-msgs==1.0.0
autoware-auto-planning-msgs==1.0.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
bloom==0.12.0
bond==4.1.2
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudpickle==1.6.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-clean==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.1
configobj==5.0.6
contourpy==1.2.1
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cvxopt==1.3.2
cycler==0.12.1
dbus-python==1.2.18
dcsl-f1tenth==0.0.0
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.5
Deprecated==1.2.13
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
duplicity==0.8.21
dwb-msgs==1.1.18
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
-e git+https://github.com/f1tenth/f1tenth_gym@cd56335eda43ff4e401331c461877227474a3ed4#egg=f110_gym
f1tenth-gym-ros==0.0.0
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
filelock==3.14.0
flake8==4.0.1
fonttools==4.51.0
fs==2.4.12
fsspec==2024.3.1
future==0.18.2
gast==0.5.2
gazebo-msgs==3.9.0
geometry-msgs==4.8.0
gpg==1.16.0
gpytorch==1.11
gym==0.19.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
interactive-markers==2.3.2
jax==0.6.0
jax-cuda12-pjrt==0.6.0
jax-cuda12-plugin==0.6.0
jaxlib==0.6.0
jaxtyping==0.2.28
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.0
joint-state-publisher==2.4.0
joy-teleop==0.3.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
linear-operator==0.5.2
llvmlite==0.43.0
lockfile==0.12.2
logging-demo==0.20.5
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
Mako==1.1.3
map-msgs==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mercurial==6.1.1
message-filters==4.3.6
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
nav-2d-msgs==1.1.18
nav-msgs==4.8.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nav2-simple-commander==1.0.0
netifaces==0.11.0
networkx==3.3
nose2==0.9.2
notify2==0.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.9.41
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.9.0.52
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
osrf-pycommon==2.1.6
packaging==24.0
paramiko==2.9.3
pathspec==0.12.1
patiencediff==0.2.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.8.0
pickle5==0.0.11
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
pyglet==1.4.11
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
PyOpenGL==3.1.9
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyserial==3.5
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
resource-retriever==3.1.3
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rospkg-modules==1.6.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.4
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scantree==0.0.1
scikit-learn==1.4.2
scipy==1.11.4
screen-resolution-extra==0.0.0
scripts==3.9.0
SecretStorage==3.3.1
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six==1.16.0
slam-toolbox==2.6.9
smclib==4.1.2
snowballstemmer==2.2.0
soupsieve==2.3.1
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
sympy==1.12
systemd-python==234
teleop-tools-msgs==1.2.1
teleop-twist-keyboard==2.4.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
threadpoolctl==3.5.0
tk==0.1.0
toml==0.10.2
topic-monitor==0.20.5
torch==2.3.0
torchvision==0.18.0
tqdm==4.67.1
trajectory-msgs==4.8.0
transforms3d==0.4.2
triton==2.3.0
turtlesim==1.4.2
typeguard==2.13.3
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
usb-creator==0.3.7
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==4.8.0
vnoise==0.1.0
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.13.3
xacro==2.0.13
xdg==5
xkit==0.0.0
zipp==1.0.0
